+++ app/containers/Event.js
@@ -1,6 +1,6 @@
 'use strict'
 
-import {get, rsvp, checkIn, forceEnd} from '../services/poop'
+import {get, rsvp, checkIn, forceEnd} from '../services/events'
 import moment             from 'moment'
 import React, {Component} from 'react'
 import {connect}          from 'react-redux'
